# public/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "ImoPanorama API",
    "version": "1.0.0",
    "description": "API documentation for ImoPanorama - Real Estate Platform in Madagascar",
    "contact": {
      "name": "ImoPanorama Support",
      "email": "contact@imopanorama.mg"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token from Better Auth"
      },
      "CookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "better-auth.session_token",
        "description": "Session cookie from Better Auth"
      }
    },
    "schemas": {
...
      "PaginatedResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "data": {
            "type": "array",
            "items": {}
          },
          "total": {
            "type": "integer",
            "description": "Total number of items"
          },
          "page": {
            "type": "integer",
            "description": "Current page number"
          },
          "totalPages": {
            "type": "integer",
            "description": "Total number of pages"
          }
        }
      },
...
      "Property": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "propertyType": {
            "type": "string",
            "enum": [
              "APARTMENT",
              "HOUSE",
              "VILLA",
              "STUDIO",
              "DUPLEX",
              "TRIPLEX",
              "PENTHOUSE",
              "LOFT",
              "TOWNHOUSE",
              "COMMERCIAL",
              "OFFICE",
              "WAREHOUSE",
              "LAND",
              "BUILDING",
              "PARKING",
              "GARAGE",
              "SHOP",
              "RESTAURANT",
              "HOTEL",
              "INDUSTRIAL",
              "AGRICULTURAL",
              "MIXED_USE",
              "VACATION_HOME",
              "OTHER"
            ]
          },
          "transactionType": {
            "type": "string",
            "enum": [
              "SALE",
              "RENT",
              "LEASE"
            ]
          },
          "price": {
            "type": "number",
            "format": "double"
          },
          "city": {
            "type": "string"
          },
          "address": {
            "type": "string",
            "nullable": true
          },
          "size": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "bedrooms": {
            "type": "integer",
            "nullable": true
          },
          "bathrooms": {
            "type": "integer",
            "nullable": true
          },
          "status": {
            "type": "string",
            "enum": [
              "AVAILABLE",
              "SOLD",
              "RENTED",
              "RESERVED",
              "UNAVAILABLE"
            ],
            "default": "AVAILABLE"
          },
          "isFeatured": {
            "type": "boolean",
            "default": false
          },
          "images": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uri"
            }
          },
          "coverImage": {
            "type": "string",
            "format": "uri",
            "nullable": true
          },
          "amenities": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "Terrain": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "price": {
            "type": "number",
            "format": "double"
          },
          "size": {
            "type": "number",
            "format": "double"
          },
          "city": {
            "type": "string"
          },
          "location": {
            "type": "string",
            "nullable": true
          },
          "type": {
            "type": "string",
            "enum": [
              "RESIDENTIAL",
              "COMMERCIAL",
              "AGRICULTURAL",
              "INDUSTRIAL",
              "MIXED",
              "OTHER"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "AVAILABLE",
              "SOLD",
              "RESERVED",
              "UNAVAILABLE"
            ],
            "default": "AVAILABLE"
          },
          "images": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uri"
            }
          },
          "coverImage": {
            "type": "string",
            "format": "uri",
            "nullable": true
          },
          "features": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "isFeatured": {
            "type": "boolean",
            "default": false
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "News": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "content": {
            "type": "string",
            "description": "HTML content"
          },
          "excerpt": {
            "type": "string",
            "nullable": true
          },
          "category": {
            "type": "string",
            "enum": [
              "ACTUALITE",
              "CONSEIL",
              "MARCHE",
              "PROJET",
              "EVENEMENT",
              "AUTRE"
            ]
          },
          "coverImage": {
            "type": "string",
            "format": "uri",
            "nullable": true
          },
          "images": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uri"
            }
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "isPublished": {
            "type": "boolean",
            "default": false
          },
          "publishedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    }
  },
...
  "paths": {
...
    "/api/admin/images": {
      "get": {
        "summary": "List all images (Admin)",
        "description": "Get a list of all uploaded images (Admin only)",
        "tags": [
          "Media"
        ],
        "security": [
          {
            "CookieAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 50
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of images"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
    "/api/admin/news": {
      "get": {
        "summary": "List all news (Admin)",
        "description": "Get all news articles including unpublished ones",
        "tags": [
          "News"
        ],
        "security": [
          {
            "CookieAuth": []
          },
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 20
            }
          },
          {
            "in": "query",
            "name": "category",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "search",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "isPublished",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of news articles",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/PaginatedResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/News"
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      },
      "post": {
        "summary": "Create news article",
        "description": "Create a new news article (Admin only)",
        "tags": [
          "News"
        ],
        "security": [
          {
            "CookieAuth": []
          },
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/News"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "News created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/api/admin/news/{id}": {
      "get": {
        "summary": "Get news by ID (Admin)",
        "tags": [
          "News"
        ],
        "security": [
          {
            "CookieAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "News article details"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "News not found"
          }
        }
      },
      "put": {
        "summary": "Update news article",
        "tags": [
          "News"
        ],
        "security": [
          {
            "CookieAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/News"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "News updated"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "News not found"
          }
        }
      },
      "delete": {
        "summary": "Delete news article",
        "tags": [
          "News"
        ],
        "security": [
          {
            "CookieAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "News deleted"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "News not found"
          }
        }
      }
    },
...
    "/api/search": {
      "get": {
        "summary": "Semantic search",
        "description": "Search properties using vector similarity with Qdrant (Public)",
        "tags": [
          "Search"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Search query",
            "example": "modern apartment near airport"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 10,
              "maximum": 50
            },
            "description": "Number of results"
          },
          {
            "in": "query",
            "name": "offset",
            "schema": {
              "type": "integer",
              "default": 0
            },
            "description": "Pagination offset"
          },
          {
            "in": "query",
            "name": "city",
            "schema": {
              "type": "string"
            },
            "description": "Filter by city"
          },
          {
            "in": "query",
            "name": "type",
            "schema": {
              "type": "string"
            },
            "description": "Filter by property type"
          },
          {
            "in": "query",
            "name": "status",
            "schema": {
              "type": "string",
              "enum": [
                "AVAILABLE",
                "SOLD",
                "RENTED",
                "RESERVED"
              ]
            },
            "description": "Filter by status"
          },
          {
            "in": "query",
            "name": "minPrice",
            "schema": {
              "type": "number"
            },
            "description": "Minimum price"
          },
          {
            "in": "query",
            "name": "maxPrice",
            "schema": {
              "type": "number"
            },
            "description": "Maximum price"
          },
          {
            "in": "query",
            "name": "minSize",
            "schema": {
              "type": "number"
            },
            "description": "Minimum size (m²)"
          },
          {
            "in": "query",
            "name": "maxSize",
            "schema": {
              "type": "number"
            },
            "description": "Maximum size (m²)"
          },
          {
            "in": "query",
            "name": "lat",
            "schema": {
              "type": "number"
            },
            "description": "Latitude for geo search"
          },
          {
            "in": "query",
            "name": "lon",
            "schema": {
              "type": "number"
            },
            "description": "Longitude for geo search"
          },
          {
            "in": "query",
            "name": "radius",
            "schema": {
              "type": "number",
              "default": 5
            },
            "description": "Search radius in km (requires lat/lon)"
          }
        ],
        "responses": {
          "200": {
            "description": "Search results with similarity scores",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "results": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "title": {
                            "type": "string"
                          },
                          "description": {
                            "type": "string"
                          },
                          "score": {
                            "type": "number",
                            "description": "Similarity score (0-1)"
                          },
                          "property": {
                            "oneOf": [
                              {
                                "$ref": "#/components/schemas/Property"
                              },
                              {
                                "$ref": "#/components/schemas/Terrain"
                              }
                            ]
                          }
                        }
                      }
                    },
                    "total": {
                      "type": "integer"
                    },
                    "query": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid query parameters"
          },
          "503": {
            "description": "Search service unavailable (fallback to database search)"
          }
        }
      }
    },
...
  }
}